Fix for_each keying of inventory
The loop through local.inventory.virtual_machine was using the loop
index as a key for the resources created by that loop. This causes the
resources to change when the order of the inventory json changes.

Instead, use the vm name as a key.

This way, we go from:

```
{
   1 => {...},
   2 => {...}
}
```

to:

```
{
  "TEST-VM-001" => {...},
  "TEST-VM-002" => {...}
}
```